[RENAME] List<Member>를 List<UserInfo>로 반환하는 Converter구현하는 메서드를 PostConverter에 위치하도록 수정
